+++ sources/EngineWorks/DBAL/Sqlite/DBAL.php
@@ -101,9 +101,14 @@ class DBAL extends AbstractDBAL
         return 'Cannot get the error because there are no active connection';
     }
 
-    protected function sqlTableEscape($tableName, $asTable)
+    public function sqlTableEscape($tableName, $asTable = '')
     {
-        return '"' . $tableName . '"' . (($asTable) ? ' AS ' . '"' . $asTable . '"' : '');
+        return '"' . $tableName . '"' . (('' !== $asTable) ? ' AS ' . '"' . $asTable . '"' : '');
+    }
+
+    public function sqlFieldEscape($fieldName, $asFieldName = '')
+    {
+        return '"' . $fieldName . '"' . (('' !== $asFieldName) ? ' AS ' . '"' . $asFieldName . '"' : '');
     }
 
     public function sqlConcatenate(...$strings)